+++ epstein_files/util/constant/names.py
@@ -184,6 +184,7 @@ BROCK_PIERCE = 'Brock Pierce'
 CHRISTIAN_EVERDELL = 'Christian Everdell'
 CHRISTOPHER_DILORIO = 'Christopher Dilorio'
 DOUGLAS_WIGDOR = 'Douglas Wigdor'
+KARYNA_SHULIAK = 'Karyna Shuliak'
 
 # No communications but name is in the files
 BILL_GATES = 'Bill Gates'